--- a/client/api_test_model.go
+++ b/client/api_test_model.go
@@ -1,10 +1,10 @@
 package client
 
 import (
-	"encoding/json"
 	"testing"
 
 	"github.com/ovn-kubernetes/libovsdb/cache"
+	"github.com/ovn-kubernetes/libovsdb/internal/json"
 	"github.com/ovn-kubernetes/libovsdb/model"
 	"github.com/ovn-kubernetes/libovsdb/ovsdb"
 	"github.com/stretchr/testify/require"