--- a/internal/json/sonic_json.go
+++ b/internal/json/sonic_json.go
@@ -1,0 +1,12 @@
+package json
+
+import (
+	"github.com/bytedance/sonic"
+)
+
+var (
+	Marshal   = sonic.Marshal
+	Unmarshal = sonic.Unmarshal
+)
+
+type RawMessage = sonic.NoCopyRawMessage