--- a/modelgen/table_test.go
+++ b/modelgen/table_test.go
@@ -1,7 +1,6 @@
 package modelgen
 
 import (
-	"encoding/json"
 	"fmt"
 	"reflect"
 	"testing"
@@ -9,6 +8,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/ovn-kubernetes/libovsdb/example/vswitchd"
+	"github.com/ovn-kubernetes/libovsdb/internal/json"
 	"github.com/ovn-kubernetes/libovsdb/model"
 	"github.com/ovn-kubernetes/libovsdb/ovsdb"
 	"github.com/stretchr/testify/assert"