--- a/ovsdb/monitor_select.go
+++ b/ovsdb/monitor_select.go
@@ -1,6 +1,6 @@
 package ovsdb
 
-import "encoding/json"
+import "github.com/ovn-kubernetes/libovsdb/internal/json"
 
 // MonitorSelect represents a monitor select according to RFC7047
 type MonitorSelect struct {