--- a/ovsdb/monitor_select_test.go
+++ b/ovsdb/monitor_select_test.go
@@ -1,8 +1,9 @@
 package ovsdb
 
 import (
-	"encoding/json"
 	"testing"
+
+	"github.com/ovn-kubernetes/libovsdb/internal/json"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"