--- a/ovsdb/named_uuid_test.go
+++ b/ovsdb/named_uuid_test.go
@@ -1,9 +1,10 @@
 package ovsdb
 
 import (
-	"encoding/json"
 	"fmt"
 	"testing"
+
+	"github.com/ovn-kubernetes/libovsdb/internal/json"
 
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"