--- a/ovsdb/update3.go
+++ b/ovsdb/update3.go
@@ -1,8 +1,9 @@
 package ovsdb
 
 import (
-	"encoding/json"
 	"fmt"
+
+	"github.com/ovn-kubernetes/libovsdb/internal/json"
 )
 
 type MonitorCondSinceReply struct {