@vtolstov 2025-08-06
please try json/v2 from std it really have near the same perf as sonic