@halfcrazy 2025-08-27
Now use sonic as the default.